- refactor advertisement and user deletion methods to return resource directly - add contacts field to user factory - update price field type in advertisements migration - remove unnecessary blank line in api routes

+++ database/migrations/2025_05_05_191632_create_advertisements_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->id();
             $table->string("name", 40);
             $table->string("description", 500);
-            $table->float("price");
+            $table->bigInteger("price");
             $table->date("created_at");
             $table->foreignId("user_id")->constrained()->cascadeOnDelete()->cascadeOnUpdate();
         });
